+++ application/Menu.java
@@ -1,0 +1,5 @@
+package application;
+
+public class Menu {
+
+}
