+++ dao/DBConnection.java
@@ -1,0 +1,5 @@
+package dao;
+
+public class DBConnection {
+
+}
