+++ entity/UserLogin.java
@@ -1,0 +1,5 @@
+package entity;
+
+public class UserLogin {
+
+}
