--- src/application/Application.java
@@ -1,8 +1,0 @@
-package application;
-
-public class Application {
-	public static void main(String[] args) {
-		Menu menu = new Menu();
-		menu.loginScreen();
-	}
-}
