--- src/dao/userLoginDAO.java
@@ -1,12 +1,0 @@
-package dao;
-
-import java.sql.Connection;
-
-public class userLoginDAO {
-	
-	private Connection connection;
-	
-	public UserLoginDAO() {
-		connection = DBConnection.getConnection();
-	}
-}
