--- src/entity/UserLogin.java
@@ -1,10 +1,0 @@
-package entity;
-
-public class UserLogin {
-
-	// private variables
-	
-	// Constructor
-	
-	//getters and setters
-}
